requirements.txt
Flask
gunicorn
python-multipart
tensorflow-cpu
opencv-python
Pillow